Switch to libvips/sharp for thumbnailing
It’s much, much faster than spawning GraphicsMagick.

The reason it was temporarily replaced by gm was that it kept crashing,
but I implemented a horrible stopgap, so it can be used in development.
With any luck, it will be fixed by the time this is deployed. (If not,
at least it can be investigated properly then.)